pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codeexporter"
version = "0.0.0"
description = "A powerful CLI tool to export project code for AI analysis."
keywords = ["code analysis", "AI", "export"]
urls = { Homepage = "https://github.com/cidraljunior/codeexporter" }
authors = [
    { name = "Aluizio Cidral Junior", email = "cidraljunior@hotmail.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]
dependencies = [
    "click>=8.0.0",
    "tqdm>=4.0.0",
    "pyyaml>=6.0.0",
    "chardet>=5.0.0",
    "pathspec>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.900",
    "isort>=5.0.0"
]

[project.scripts]
codeexport = "codeexporter.cli:main"

[tool.setuptools]
packages = ["codeexporter"]
package-dir = {"" = "."}

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.0.0"  # Initial version
version_files = [
    "pyproject.toml:version"
]
tag_format = "v$version"